+++ packages/build-tools/src/steps/functions/deploy.ts
@@ -1,10 +1,10 @@
 import { UserError } from '@expo/eas-build-job';
+import { PipeMode, bunyan } from '@expo/logger';
 import {
   BuildFunction,
   BuildStepInput,
   BuildStepInputValueTypeName,
   BuildStepOutput,
-  spawnAsync,
 } from '@expo/steps';
 
 import { runEasCliCommand } from '../../utils/easCli';
@@ -31,11 +31,12 @@ export function createEasDeployBuildFunction(): BuildFunction {
         allowedValueTypeName: BuildStepInputValueTypeName.BOOLEAN,
         required: false,
       }),
+      // Match eas/export default `output_dir`
       BuildStepInput.createProvider({
-        id: 'export_command',
+        id: 'export_dir',
         allowedValueTypeName: BuildStepInputValueTypeName.STRING,
         required: false,
-        defaultValue: 'npx expo export --platform web',
+        defaultValue: 'dist',
       }),
     ],
     outputProviders: [
@@ -43,29 +44,35 @@ export function createEasDeployBuildFunction(): BuildFunction {
         id: 'deploy_json',
         required: true,
       }),
+      BuildStepOutput.createProvider({
+        id: 'deploy_url',
+        required: false,
+      }),
+      BuildStepOutput.createProvider({
+        id: 'deploy_deployment_url',
+        required: false,
+      }),
+      BuildStepOutput.createProvider({
+        id: 'deploy_identifier',
+        required: false,
+      }),
+      BuildStepOutput.createProvider({
+        id: 'deploy_dashboard_url',
+        required: false,
+      }),
+      BuildStepOutput.createProvider({
+        id: 'deploy_alias_url',
+        required: false,
+      }),
     ],
     fn: async (stepsCtx, { inputs, outputs, env }) => {
       const alias = inputs.alias.value as string | undefined;
       const prod = inputs.prod.value as boolean | undefined;
       const sourceMaps = inputs.source_maps.value as boolean | undefined;
-      const exportCommand = inputs.export_command.value as string;
-
-      stepsCtx.logger.info(`Running export command: ${exportCommand}`);
-      try {
-        await spawnAsync('sh', ['-c', exportCommand], {
-          cwd: stepsCtx.workingDirectory,
-          env,
-          logger: stepsCtx.logger,
-          stdio: 'pipe',
-        });
-      } catch (error) {
-        throw new UserError(
-          'EAS_DEPLOY_EXPORT_FAILED',
-          `Export command failed: ${error instanceof Error ? error.message : 'unknown error'}`
-        );
-      }
+      const exportDir = inputs.export_dir.value as string;
 
       const deployCommand = getDeployCommand({
+        exportDir,
         alias,
         prod,
         environment: stepsCtx.global.staticContext.metadata?.environment,
@@ -80,9 +87,21 @@ export function createEasDeployBuildFunction(): BuildFunction {
             cwd: stepsCtx.workingDirectory,
             env,
             logger: stepsCtx.logger,
+            mode: PipeMode.STDERR_ONLY_AS_STDOUT,
           },
         });
-        outputs.deploy_json.set(result.stdout.toString());
+
+        const deployJson = result.stdout.toString();
+        outputs.deploy_json.set(deployJson);
+        const parsedDeploymentOutput = parseDeploymentOutput({
+          deployJson,
+          logger: stepsCtx.logger,
+        });
+        if (parsedDeploymentOutput) {
+          for (const [key, value] of Object.entries(parsedDeploymentOutput)) {
+            outputs[key].set(value);
+          }
+        }
       } catch (error) {
         throw new UserError(
           'EAS_DEPLOY_FAILED',
@@ -94,33 +113,59 @@ export function createEasDeployBuildFunction(): BuildFunction {
 }
 
 function getDeployCommand({
+  exportDir,
   alias,
   prod,
   environment,
   sourceMaps,
 }: {
+  exportDir: string;
   alias?: string;
   prod?: boolean;
   environment?: string;
   sourceMaps?: boolean;
 }): string[] {
-  const deployCommand = ['deploy', '--non-interactive', '--json'];
-
+  const deployCommand = ['deploy', '--non-interactive', '--json', '--export-dir', exportDir];
   if (environment) {
     deployCommand.push('--environment', environment);
   }
-
   if (alias) {
     deployCommand.push('--alias', alias);
   }
-
   if (prod) {
     deployCommand.push('--prod');
   }
-
   if (sourceMaps) {
     deployCommand.push('--source-maps');
   }
-
   return deployCommand;
+}
+
+function parseDeploymentOutput({ deployJson, logger }: { deployJson: string; logger: bunyan }): {
+  deploy_url: string;
+  deploy_deployment_url: string;
+  deploy_identifier: string;
+  deploy_dashboard_url: string;
+  deploy_alias_url: string;
+} | null {
+  try {
+    // TODO: improve typing here; Look into WorkerDeploymentData
+    const deployObject = JSON.parse(deployJson);
+    return {
+      deploy_url: deployObject.production.url || deployObject.aliases[0].url || deployObject.url,
+      deploy_deployment_url: deployObject.url,
+      deploy_identifier: deployObject.identifier,
+      deploy_dashboard_url: deployObject.dashboardUrl,
+      deploy_alias_url: deployObject.aliases[0].url,
+    };
+  } catch (error: unknown) {
+    const message = error instanceof Error ? error.message : String(error);
+    logger.warn(
+      'Failed to parse deploy JSON: ' +
+        `${message}\n` +
+        'DEPLOY_URL, DEPLOY_DEPLOYMENT_URL, DEPLOY_IDENTIFIER, DEPLOY_DASHBOARD_URL, DEPLOY_ALIAS_URL will be unavailable on this build.',
+      { error }
+    );
+    return null;
+  }
 }